feat: add operator detect func & add manifest
